+++ XMLRead.cpp
@@ -143,14 +143,14 @@ std::vector<attribute> XMLRead::_GetAttributes(std::string fieldName) {
 
 bool XMLRead::_CreateChild(std::string name, std::string content, std::string parentTagName) {
     std::cout << "Name = " << name << ", content = " << content << ", parentTagName = " << parentTagName << std::endl;
-    unsigned int loc = (unsigned)_xmlData.find(_MakeTag(parentTagName, true)) + parentTagName.length() + 2;
+    size_t loc = _xmlData.find(_MakeTag(parentTagName, true)) + parentTagName.length() + 2;
     _xmlData.insert(loc, _MakeTag(name, true) + content + _MakeTag(name, false));
 
     return true;
 }
 
 void XMLRead::_DeleteAttributes(std::string fieldName) {
-    unsigned int loc = (unsigned)_xmlData.find("<" + fieldName) + fieldName.length();
+    size_t loc = _xmlData.find("<" + fieldName) + fieldName.length();
     std::string attributeContents = _GetStringBetween(_xmlData, "<" + fieldName, ">");
     if (attributeContents.length() > 0) {
         std::cout << "ATT LENGTH = " << attributeContents.length() << "|";
